ant_plus_sensor/python_ant_plus_sensor/pyproject.toml
[tool.poetry]
name = "python_ant_plus_sensor"
version = "0.1.0"
description = "Communicate with ANT+ devices using Python ANT+ library"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
paho-mqtt = "^1.6.1"
pyusb = "^1.2.1"
libusb = "^1.0.26b5"
openant = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
